+++ src/core/mod.rs
@@ -1,3 +1,6 @@
 pub mod event;
 pub mod schedule_stream;
 pub mod transposer;
+
+pub use event::Event;
+pub use transposer::Transposer;
