+++ src/core/schedule_stream/mod.rs
@@ -1,4 +1,12 @@
-pub mod realtime_stream;
-pub mod schedule_stream;
-pub mod schedule_stream_ext;
-pub mod timestamp;
+mod realtime_stream;
+mod schedule_stream;
+mod schedule_stream_ext;
+mod timestamp;
+
+pub use schedule_stream::*;
+
+pub use realtime_stream::RealtimeStream;
+
+pub use schedule_stream_ext::ScheduleStreamExt;
+
+pub use timestamp::Timestamp;
