+++ src/core/transposer/mod.rs
@@ -1,10 +1,15 @@
-pub mod transposer;
-pub mod transposer_context;
-pub mod transposer_engine;
-pub mod transposer_event;
-pub mod transposer_expire_handle;
+mod context;
+mod engine;
+mod transposer;
 
-pub(self) mod transposer_engine_internal;
+pub mod transposer_event;
+pub use context::TransposerContext;
+pub use engine::TransposerEngine;
+pub use expire_handle::ExpireHandle;
+pub use transposer::*;
+
+pub(self) mod engine_internal;
+pub(self) mod expire_handle;
 pub(self) mod transposer_frame;
 pub(self) mod transposer_function_wrappers;
 pub(self) mod transposer_update;
