+++ src/utilities/mod.rs
@@ -1,6 +1,2 @@
-pub mod debug_sink;
-pub mod debug_waker;
 pub mod full_ord;
 pub mod winit;
-pub mod debug_future;
-pub mod debug_stream;
